+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased - 97
 
 - Fixed pressing the save button multiple times creating multiple entries
+- Lower card header size when hiding details to fit even more cards
 
 ## v2.12.0 - 96 (2021-12-23)